requirements.txt
numpy==1.22.0
tensorflow==2.8.0
keras-tuner==1.1.0
python-chess==0.31.3
